Move NcaUtils
