+++ src/LibHac/FsSrv/FsCreator/StorageOnNcaCreator.cs
@@ -5,7 +5,7 @@ using LibHac.Fs;
 using LibHac.Fs.Fsa;
 using LibHac.FsSystem;
 using LibHac.FsSystem.Impl;
-using LibHac.FsSystem.NcaUtils;
+using LibHac.Tools.FsSystem.NcaUtils;
 
 namespace LibHac.FsSrv.FsCreator;
