+++ src/LibHac/FsSrv/NcaFileSystemServiceImpl.cs
@@ -8,10 +8,10 @@ using LibHac.Fs.Fsa;
 using LibHac.FsSrv.FsCreator;
 using LibHac.FsSrv.Impl;
 using LibHac.FsSystem;
-using LibHac.FsSystem.NcaUtils;
 using LibHac.Ncm;
 using LibHac.Os;
 using LibHac.Spl;
+using LibHac.Tools.FsSystem.NcaUtils;
 using LibHac.Util;
 using RightsId = LibHac.Fs.RightsId;
 using Utility = LibHac.FsSystem.Utility;
