+++ src/LibHac/Tools/FsSystem/NcaUtils/NcaFsIntegrityInfoIvfc.cs
@@ -2,7 +2,7 @@
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-namespace LibHac.FsSystem.NcaUtils;
+namespace LibHac.Tools.FsSystem.NcaUtils;
 
 public struct NcaFsIntegrityInfoIvfc
 {
