+++ src/hactoolnet/ProcessDelta.cs
@@ -6,8 +6,8 @@ using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
 using LibHac.FsSystem;
-using LibHac.FsSystem.NcaUtils;
 using LibHac.Tools.FsSystem;
+using LibHac.Tools.FsSystem.NcaUtils;
 using static hactoolnet.Print;
 
 namespace hactoolnet;
