+++ src/hactoolnet/ProcessSwitchFs.cs
@@ -8,7 +8,6 @@ using LibHac;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
 using LibHac.FsSystem;
-using LibHac.FsSystem.NcaUtils;
 using LibHac.Ns;
 using LibHac.Tools.Fs;
 using LibHac.Tools.FsSystem.Save;
